requirements.txt
aiofile==3.8.8
aiohttp==3.8.6
aiosignal==1.3.1
amazon-braket-default-simulator==1.20.1
amazon-braket-schemas==1.19.1.post0
amazon-braket-sdk==1.59.2
annotated-types==0.6.0
antlr4-python3-runtime==4.9.2
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
azure-core==1.29.5
azure-identity==1.15.0
azure-quantum==0.28.302812
azure-storage-blob==12.18.3
backoff==2.2.1
boltons==23.0.0
boto3==1.28.73
botocore==1.31.73
caio==0.9.13
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
cloudpickle==2.2.1
contourpy==1.1.1
cryptography==41.0.5
cycler==0.12.1
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.7
fastapi==0.104.0
fonttools==4.43.1
frozenlist==1.4.0
general-superstaq==0.4.29
gql==3.4.1
graphql-core==3.2.3
h11==0.14.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
idna==3.4
importlib-metadata==6.8.0
isodate==0.6.1
jmespath==1.0.1
kiwisolver==1.4.5
Markdown==3.5
matplotlib==3.8.0
mpmath==1.3.0
msal==1.24.1
msal-extensions==1.0.0
msrest==0.7.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
ntlm-auth==1.5.0
numpy==1.23.5
oauthlib==3.2.2
openpulse==0.4.2
openqasm3==0.4.0
opt-einsum==3.3.0
oqpy==0.2.1
packaging==23.2
pbr==5.11.1
Pillow==10.1.0
platformdirs==3.11.0
ply==3.11
portalocker==2.8.2
psutil==5.9.6
py==1.11.0
pycparser==2.21
pydantic==1.10.13
pydantic_core==2.10.1
PyJWT==2.8.0
pylatexenc==2.10
pyparsing==3.1.1
pyqir==0.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-markdown-math==0.8
PyYAML==6.0.1
qiskit==0.44.3
qiskit-aer==0.11.2
qiskit-braket-provider==0.0.4
qiskit-ibmq-provider==0.19.2
qiskit-ionq==0.4.3
qiskit-qir==0.3.1
qiskit-terra==0.25.3
qubovert==1.2.5
requests==2.31.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
retry==0.9.2
retrying==1.3.3
rustworkx==0.13.2
s3transfer==0.7.0
scipy==1.11.3
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
stevedore==5.1.0
strangeworks==0.4.10
strangeworks-core==0.2.7
symengine==0.9.2
sympy==1.12
tabulate==0.9.0
tomlkit==0.11.8
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.6.4
websockets==12.0
wrapt==1.15.0
yarl==1.9.2
zipp==3.17.0
